requirements.txt
faiss-cpu==1.10.0
fastapi==0.115.11
langchain-openai==0.3.8
numpy==1.26.4
openai==1.61.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
SQLAlchemy==2.0.38
sqlite-utils==3.38
streamlit==1.43.2
uvicorn==0.34.0
